Introduce operator compositions P(Q)
This represents the application of P on the operator Q, in
general not equal to PQ. If P and Q are both QMPotentials,
then they will be combined into a single potential. If P
is a derivative and Q is a potential, then P(Q) results in
the differentiated potential and the derivative does NOT
carry through by the chain rule.